Initial base setup
To save time for the sessions I've set up an initial simple Sinatra app complete
with basic test environment ready with "hello world" examples in Cucumber and
RSpec to prove the setup works fully.

I've gone for SQLite as a simple database with less dependencies needed and
also included the gems we may as well use to complete the app.

For better or worse I've used binstubs to try and keep everyone's installations
on the same page.